Isolate video stabilization from engine facade
The stabilization workflow is self-contained enough to leave engine.py while preserving mcp_video.engine.stabilize for public callers. The extracted module keeps the vidstab two-pass flow and uses shared timeout, quality, numeric sanitation, and filter escaping helpers.

Constraint: Preserve existing vidstab availability behavior while continuing engine decomposition.

Rejected: Combine stabilize with masking | both are output transforms but use different filters, validations, and failure modes.

Confidence: high

Scope-risk: narrow

Directive: Keep vidstab detect/transform behavior covered before changing filter availability checks or vector-file handling.

Tested: ruff check mcp_video/engine.py mcp_video/engine_stabilize.py; ruff format --check mcp_video/engine.py mcp_video/engine_stabilize.py; python3 -m pytest tests/test_engine_advanced.py -k 'stabilize' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'stabilize' tests/test_server.py::TestVideoStabilizeTool tests/test_red_team.py -k 'stabilize' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
